chyba ostatnio to co Ci w wordzie posłałem było już po tym jak grzebałęm w kodzie i namieszałem . teraz znalazłęm w gicie zmieny z 13.03  wprowadziłem je jeszcze raz i add na Client działął i na podstawie tego zmieniłęm choucha i treaning ... treaning dlaej ma problem ale nie wiem jak wprowadzić do niego dane do BirthDate :-) reszta powinna być ok

+++ src/main/java/org/szpax/brothers/controller/ClientController.java
@@ -34,12 +34,12 @@ public class ClientController {
     }
 
     @RequestMapping("/{clientId}")
-    public Client one(@PathVariable Long clientId) {
+    public Client one(@PathVariable Integer clientId) {
         return clientService.findClientById(clientId);
     }
 
     @RequestMapping("/{clientId}/update")
-    public Client update(@PathVariable Long clientId,
+    public Client update(@PathVariable Integer clientId,
                          @RequestParam(required = false) String firstName,
                          @RequestParam(required = false) String lastName) {
         return clientService.update(clientId, firstName, lastName);
